Added founders section to FR
Flowmark did a lot of funny stuff breaking the year off and adding a new line and unindenting some citations. I corrected it so it looks right in Markdown Preview but a bit confused. It's seems like Flowmark that's adding the \ after the year and breaking the lines. I manually added \ to years without them for consistency.