AMT updates
There have been some major 2018 changes to threshholds and exemptions on AMT. I have updated the income threshold in the text and updated two out-of-date links but we should revisit this. While no one wants to be hit with the AMT it looks like recent legislation is making it apply to fewer and fewer, warranting discussion in the text and possibly less severe warninsg.